refactor(renderer): 폴링 헬퍼를 공용 훅으로 분리
usePolledValue가 useLiveAiSessions 안에만 있어 다른 화면이 같은 폴링 규칙
(앞선 조회 미완료 시 건너뛰기, 창 비활성 시 정지)을 쓰려면 복사해야 했다.
동작은 그대로 두고 자리만 옮긴다.
